Rework how popups get data.
Too much information was being provided in the 'uri' for each resource which caused some records with 30+ resource to fail to load the popup. An error, 414 Request URI too large, was being returned. This update reduces the length of the URI by moving part of the process to the API.