Fix bug with existing email addresses
If a reviewer was given using an email address, and a user already
existed with that email address no one would be notified. Now there
is a check to see if the email is for a new user or not. If the user
already exists, they should receive an email. Otherwise, new reviewers
will only get their invitation email.