Update how reviewers are stored
This allows greater delination between reviewers and othere users who have
the same email address. Both email and user name are stored in the appropriate fields:
`maintainer` and `maintainer_email`. Otherwise, an admin or Sysadmin account was
being treated like a reviewer when they shared email addresses.

Probably was never going to be an issue, but why not.