Add second search box on results page
Previously the only search box was part of the nav bar. This search
is in a differnt <form> from the sorting options. This meant that they
acted independently. The new search box is in the same <form> as the
sort options which allows the user to sort the result of the search.